Modify auth to get token from user object
